requirementx.txt
Django~=3.1.7
PyAutoGUI~=0.9.52
gunicorn ~=20.1.0
docker
psycopg2
validate-email
djangorestframework
django-rest-swagger
drf_yasg
docker
djangorestframework-simplejwt